Fix some minor things.

+++ shared/configs/arguments.py
@@ -55,18 +55,18 @@ class SplitConf:
 
 @dataclass
 class DataModuleConf:
-    # DataLoader settings
     batch_size_tr: int = 1
     batch_size_dep: Optional[int] = None
     batch_size_te: Optional[int] = None
     num_samples_per_group_per_bag: int = 1
-
     num_workers: int = 0
     persist_workers: bool = False
     pin_memory: bool = True
     gt_deployment: bool = True
+    # Amount of noise to apply to the labels used for balanced sampling
+    # -- only applicable when ``gt_deployment=True``
     label_noise: float = 0.0
-    seed: float = 47
+    seed: int = 47
 
 
 @dataclass
